merge: apply latest stashed code from feat/owner-mypg-new-pg-form
Merged 172 files of uncommitted work that was stashed locally.
This includes all latest production code from the feature branch.

Key changes:
- Applied all latest mypg feature code (new_pg_setup_screen, forms, widgets)
- Merged updated common widgets (adaptive_app_bar, primary_button, text_input)
- Updated models, repositories, viewmodels across owner/guest dashboards
- Preserved macOS build fixes and security improvements
- Excluded .cursor/mcp.json (contains secrets, already in .gitignore)

+++ lib/feature/owner_dashboard/mypg/data/models/pg_floor_model.dart
@@ -28,7 +28,7 @@ class PgFloorModel {
     return PgFloorModel(
       floorId: map['floorId'] ?? '',
       floorName: map['floorName'] ?? '',
-      floorNumber: map['floorNumber'] ?? 0,
+      floorNumber: (map['floorNumber'] ?? 0).toInt(),
       rooms: (map['rooms'] as List<dynamic>?)
               ?.map((r) => PgRoomModel.fromMap(r as Map<String, dynamic>))
               .toList() ??
@@ -82,4 +82,3 @@ class PgFloorModel {
     );
   }
 }
-
